pyproject.toml
[project]
name = "mermaids"
version = "0.1.0"
description = "Telescope mount control and ZWO ASI camera driver for astronomical imaging"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["astronomy", "telescope", "mount", "zwo", "asi", "camera", "astrophotography", "gps"]

dependencies = [
    # Mount driver dependencies
    "plotly>=5.0.0",
    "pyserial>=3.5",
    "pynmea2>=1.19.0",

    # ASI camera driver dependencies
    "astropy>=6.0.1",
    "fitsio>=1.2.0",
    "matplotlib>=3.9.4",
    "numcodecs>=0.12.0",
    "numpy>=1.26.4",
    "pillow>=11.2.1",
    "scipy>=1.13.1",
    "tqdm>=4.67.1",
    "zarr>=2.18.0",
    "zwoasi>=0.2.0",
    "tetra3 @ git+https://github.com/esa/tetra3.git",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "jupyter>=1.1.1",
    "line-profiler>=5.0.0",
    "pytest>=7.0.0",
]

[project.scripts]
# ASI camera CLI commands
asi-focus = "asi_driver.cli.focus:main"
asi-burst = "asi_driver.cli.burst:main"
asi-gps-test = "asi_driver.cli.gps_test:main"
asi-cam-setup = "asi_driver.cli.cam_setup:main"

# Mount driver CLI commands
mount-single = "mount_driver.cli.mount:main"
mount-multi = "mount_driver.cli.mounts:main"
mount-calibrate = "mount_driver.cli.calibrate:main"
mount-diagnose = "mount_driver.cli.diagnose:main"
mount-observe = "mount_driver.cli.observe:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["asi_driver/src/asi_driver", "src/mount_driver"]
